:label:Rename: 리뷰 기능별 파일 분리 #31
1. review.js에는 리뷰 조회 및 삭제 함수가 있습니다.
2. review-posting.js에는 리뷰 작성 및 관련 함수가 있습니다.
3. review-editing.js에는 리뷰 수정 및 관련 함수가 있습니다.
